Pull common components of testchip_tsi/testchip_dtm into testchip_htif_t
