+++ src/main/resources/testchipip/csrc/testchip_tsi.h
@@ -5,14 +5,9 @@
 
 #include <fesvr/tsi.h>
 #include <fesvr/htif.h>
+#include "testchip_htif.h"
 
-struct init_access_t {
-  uint64_t address;
-  uint32_t stdata;
-  bool store;
-};
-
-class testchip_tsi_t : public tsi_t
+class testchip_tsi_t : public tsi_t, public testchip_htif_t
 {
  public:
   testchip_tsi_t(int argc, char** argv, bool has_loadmem);
@@ -37,7 +32,5 @@ class testchip_tsi_t : public tsi_t
  private:
 
   bool is_loadmem;
-  bool write_hart0_msip;
-  std::vector<init_access_t> init_accesses;
 };
 #endif
